form_centre-portal/requirements.txt
# Browser Automation
selenium==4.18.1
webdriver-manager==4.0.1

# Google Sheets
gspread==6.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0

# Excel fallback
openpyxl==3.1.2

# Environment variables
python-dotenv==1.0.1

# Utilities
pandas==2.2.1
retry==0.9.2
colorlog==6.8.2

# ── Captcha Solving ────────────────────────────────
openai-whisper
torch
SpeechRecognition==3.10.1
pydub==0.25.1
requests==2.31.0
opencv-python
pytesseract